+++ internal/xerrors/transport_test.go
@@ -1,152 +1,126 @@
 package xerrors
 
 import (
-	"context"
-	"errors"
 	"fmt"
 	"testing"
 
 	"github.com/stretchr/testify/require"
-	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb"
 	grpcCodes "google.golang.org/grpc/codes"
 	grpcStatus "google.golang.org/grpc/status"
+
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xtest"
 )
 
 func TestIsTransportError(t *testing.T) {
-	code := grpcCodes.Canceled
-	for _, err := range []error{
-		&transportError{status: grpcStatus.New(code, "")},
-		fmt.Errorf("wrapped: %w", &transportError{status: grpcStatus.New(code, "")}),
-	} {
-		t.Run("", func(t *testing.T) {
-			if !IsTransportError(err, code) {
-				t.Errorf("expected %v to be transportError with code=%v", err, code)
-			}
-		})
-	}
-}
-
-func TestIsNonTransportError(t *testing.T) {
-	code := grpcCodes.Canceled
-	for _, err := range []error{
-		&transportError{status: grpcStatus.New(grpcCodes.Aborted, "")},
-		fmt.Errorf("wrapped: %w", &transportError{status: grpcStatus.New(grpcCodes.Aborted, "")}),
-		&operationError{code: Ydb.StatusIds_BAD_REQUEST},
-	} {
-		t.Run("", func(t *testing.T) {
-			if IsTransportError(err, code) {
-				t.Errorf("expected %v not to be transportError with code=%v", err, code)
-			}
-		})
-	}
-}
-
-func TestIsNonOperationError(t *testing.T) {
-	code := Ydb.StatusIds_BAD_REQUEST
-	for _, err := range []error{
-		&operationError{code: Ydb.StatusIds_TIMEOUT},
-		fmt.Errorf("wrapped: %w", &operationError{code: Ydb.StatusIds_TIMEOUT}),
-		&transportError{status: grpcStatus.New(grpcCodes.Aborted, "")},
-	} {
-		t.Run("", func(t *testing.T) {
-			if IsOperationError(err, code) {
-				t.Errorf("expected %v not to be operationError with code=%v", err, code)
-			}
-		})
-	}
-}
-
-func TestMustPessimizeEndpoint(t *testing.T) {
-	for _, test := range []struct {
-		error     error
-		pessimize bool
+	for _, tt := range []struct {
+		name  string
+		err   error
+		codes []grpcCodes.Code
+		match bool
 	}{
+		// check only transport error with any grpc status code
 		{
-			error:     Transport(grpcStatus.Error(grpcCodes.Canceled, "")),
-			pessimize: true,
+			name:  xtest.CurrentFileLine(),
+			err:   grpcStatus.Error(grpcCodes.Canceled, ""),
+			match: true,
 		},
 		{
-			error:     Transport(grpcStatus.Error(grpcCodes.Unknown, "")),
-			pessimize: true,
+			name:  xtest.CurrentFileLine(),
+			err:   &transportError{status: grpcStatus.New(grpcCodes.Canceled, "")},
+			match: true,
 		},
 		{
-			error:     Transport(grpcStatus.Error(grpcCodes.InvalidArgument, "")),
-			pessimize: true,
+			name:  xtest.CurrentFileLine(),
+			err:   fmt.Errorf("wrapped: %w", &transportError{status: grpcStatus.New(grpcCodes.Canceled, "")}),
+			match: true,
 		},
 		{
-			error:     Transport(grpcStatus.Error(grpcCodes.DeadlineExceeded, "")),
-			pessimize: true,
+			name:  xtest.CurrentFileLine(),
+			err:   fmt.Errorf("wrapped: %w", grpcStatus.Error(grpcCodes.Canceled, "")),
+			match: true,
 		},
 		{
-			error:     Transport(grpcStatus.Error(grpcCodes.NotFound, "")),
-			pessimize: true,
+			name: xtest.CurrentFileLine(),
+			err: Join(
+				fmt.Errorf("test"),
+				grpcStatus.Error(grpcCodes.Canceled, ""),
+				Retryable(fmt.Errorf("test")),
+			),
+			match: true,
+		},
+		// match grpc status code
+		{
+			name:  xtest.CurrentFileLine(),
+			err:   grpcStatus.Error(grpcCodes.Canceled, ""),
+			codes: []grpcCodes.Code{grpcCodes.Canceled},
+			match: true,
 		},
 		{
-			error:     Transport(grpcStatus.Error(grpcCodes.AlreadyExists, "")),
-			pessimize: true,
+			name:  xtest.CurrentFileLine(),
+			err:   &transportError{status: grpcStatus.New(grpcCodes.Canceled, "")},
+			codes: []grpcCodes.Code{grpcCodes.Canceled},
+			match: true,
 		},
 		{
-			error:     Transport(grpcStatus.Error(grpcCodes.PermissionDenied, "")),
-			pessimize: true,
+			name:  xtest.CurrentFileLine(),
+			err:   fmt.Errorf("wrapped: %w", &transportError{status: grpcStatus.New(grpcCodes.Canceled, "")}),
+			codes: []grpcCodes.Code{grpcCodes.Canceled},
+			match: true,
 		},
 		{
-			error:     Transport(grpcStatus.Error(grpcCodes.ResourceExhausted, "")),
-			pessimize: false,
+			name:  xtest.CurrentFileLine(),
+			err:   fmt.Errorf("wrapped: %w", grpcStatus.Error(grpcCodes.Canceled, "")),
+			codes: []grpcCodes.Code{grpcCodes.Canceled},
+			match: true,
 		},
 		{
-			error:     Transport(grpcStatus.Error(grpcCodes.FailedPrecondition, "")),
-			pessimize: true,
+			name: xtest.CurrentFileLine(),
+			err: Join(
+				fmt.Errorf("test"),
+				grpcStatus.Error(grpcCodes.Canceled, ""),
+				Retryable(fmt.Errorf("test")),
+			),
+			codes: []grpcCodes.Code{grpcCodes.Canceled},
+			match: true,
+		},
+		// no match grpc status code
+		{
+			name:  xtest.CurrentFileLine(),
+			err:   grpcStatus.Error(grpcCodes.Canceled, ""),
+			codes: []grpcCodes.Code{grpcCodes.Aborted},
+			match: false,
 		},
 		{
-			error:     Transport(grpcStatus.Error(grpcCodes.Aborted, "")),
-			pessimize: true,
+			name:  xtest.CurrentFileLine(),
+			err:   &transportError{status: grpcStatus.New(grpcCodes.Canceled, "")},
+			codes: []grpcCodes.Code{grpcCodes.Aborted},
+			match: false,
 		},
 		{
-			error:     Transport(grpcStatus.Error(grpcCodes.OutOfRange, "")),
-			pessimize: false,
+			name:  xtest.CurrentFileLine(),
+			err:   fmt.Errorf("wrapped: %w", &transportError{status: grpcStatus.New(grpcCodes.Canceled, "")}),
+			codes: []grpcCodes.Code{grpcCodes.Aborted},
+			match: false,
 		},
 		{
-			error:     Transport(grpcStatus.Error(grpcCodes.Unimplemented, "")),
-			pessimize: true,
+			name:  xtest.CurrentFileLine(),
+			err:   fmt.Errorf("wrapped: %w", grpcStatus.Error(grpcCodes.Canceled, "")),
+			codes: []grpcCodes.Code{grpcCodes.Aborted},
+			match: false,
 		},
 		{
-			error:     Transport(grpcStatus.Error(grpcCodes.Internal, "")),
-			pessimize: true,
-		},
-		{
-			error:     Transport(grpcStatus.Error(grpcCodes.Unavailable, "")),
-			pessimize: true,
-		},
-		{
-			error:     Transport(grpcStatus.Error(grpcCodes.DataLoss, "")),
-			pessimize: true,
-		},
-		{
-			error:     Transport(grpcStatus.Error(grpcCodes.Unauthenticated, "")),
-			pessimize: true,
-		},
-		{
-			error:     context.Canceled,
-			pessimize: false,
-		},
-		{
-			error:     context.DeadlineExceeded,
-			pessimize: false,
-		},
-		{
-			error:     fmt.Errorf("user error"),
-			pessimize: false,
+			name: xtest.CurrentFileLine(),
+			err: Join(
+				fmt.Errorf("test"),
+				grpcStatus.Error(grpcCodes.Canceled, ""),
+				Retryable(fmt.Errorf("test")),
+			),
+			codes: []grpcCodes.Code{grpcCodes.Aborted},
+			match: false,
 		},
 	} {
-		err := errors.Unwrap(test.error)
-		if err == nil {
-			err = test.error
-		}
-		t.Run(err.Error(), func(t *testing.T) {
-			pessimize := MustPessimizeEndpoint(test.error)
-			if pessimize != test.pessimize {
-				t.Errorf("unexpected pessimization status for error `%v`: %t, exp: %t", test.error, pessimize, test.pessimize)
-			}
+		t.Run(tt.name, func(t *testing.T) {
+			require.Equal(t, tt.match, IsTransportError(tt.err, tt.codes...))
 		})
 	}
 }
@@ -185,6 +159,42 @@ func Test_transportError_Error(t *testing.T) {
 	} {
 		t.Run("", func(t *testing.T) {
 			require.Equal(t, tt.text, tt.err.Error())
+		})
+	}
+}
+
+func TestTransportErrorName(t *testing.T) {
+	for _, tt := range []struct {
+		err  error
+		name string
+	}{
+		{
+			err:  nil,
+			name: "",
+		},
+		{
+			err:  grpcStatus.Error(grpcCodes.Aborted, ""),
+			name: "transport/Aborted",
+		},
+		{
+			err:  TransportError(grpcStatus.Error(grpcCodes.Aborted, "")),
+			name: "transport/Aborted",
+		},
+		{
+			err:  WithStackTrace(grpcStatus.Error(grpcCodes.Aborted, "")),
+			name: "transport/Aborted",
+		},
+		{
+			err:  WithStackTrace(TransportError(grpcStatus.Error(grpcCodes.Aborted, ""))),
+			name: "transport/Aborted",
+		},
+	} {
+		t.Run("", func(t *testing.T) {
+			if tt.err == nil {
+				require.Nil(t, TransportError(tt.err))
+			} else {
+				require.Equal(t, tt.name, TransportError(tt.err).Name())
+			}
 		})
 	}
 }